requirements.txt
aiofiles
beautifulsoup4
cryptography
fastapi
faiss-cpu
flask
flask-cors
gradio
ibm-watsonx-ai
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-ibm
langchain-text-splitters
langgraph
lxml
numpy
pandas
python-dotenv
pypdf
requests
scikit-learn
sentence-transformers
uvicorn[standard]
yfinance
gunicorn